% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ages.R
\docType{methods}
\name{ages}
\alias{ages}
\alias{ages,FLCohort-method}
\alias{ages,FLQuant-method}
\alias{ages-method}
\title{ages}
\usage{
ages(object, ...)
}
\arguments{
\item{object}{FLQuant or FLCohort}

\item{...}{any other arguments}
}
\value{
\code{FLQuant} or \code{FLCohort}
}
\description{
Creates FLQuant/FLCohort with ages as entries
}
\examples{
\dontrun{
data(ple4)
ages(m(ple4))}
}
\seealso{
\code{\link{knife}} \code{\link{gascuel}} \code{\link{sigmoid}} \code{\link{gompertz}} \code{\link{vonB}} \code{\link{dnormal}} \code{\link{logistic}}
}

