% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cas}
\alias{cas}
\title{cas}
\format{An \code{data.frame} object
\describe{
  \item{year}{ of capture}
  \item{len}{ at capture}   
  \item{n}{ frequency}   
}}
\description{
A dataset containing lengths by year
}
\seealso{
\code{\link{teleost}}
}

