\name{data}
\alias{ple4}
\alias{ple4sex}
\title{FLife datasets}
\description{
\itemize{
    \item{ple4}{A dataset of North Sea (ICES Area IV) plaice catch, yield, landings, discards, natural mortality, weight-at-age and maturity, together with the VPA estimated abundances and fishing mortalities, contained in an FLStock object.}
    \item{ple4sex}{A dataset of North Sea (ICES Area IV) plaice disaggregated by sex catch, yield, landings, discards, natural mortality, weight-at-age and maturity, together with the VPA estimated abundances and fishing mortalities, contained in an FLStock object.}
	}
}
\seealso{
\linkS4class{FLStock}, \linkS4class{FLSR}, \linkS4class{FLBiol}
}
\references{

}
\keyword{datasets}
