% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnormal.R
\docType{methods}
\name{dnormal,FLQuant,FLPar-method}
\alias{dnormal}
\alias{dnormal,FLPar,FLPar-method}
\alias{dnormal,FLQuant,FLPar-method}
\alias{dnormal,FLQuant,numeric-method}
\alias{dnormal,numeric,numeric-method}
\alias{dnormal-method}
\title{Double normal ogive}
\usage{
\S4method{dnormal}{FLQuant,FLPar}(age, params, ...)
}
\arguments{
\item{age}{FLQuant or FLCohort}

\item{params}{\code{FLPar} with parameters \code{a1} age at maximum, \code{sl} SD for lefthand limb and \code{sr} SD for righthand limb.}

\item{...}{any other arguments}
}
\value{
Returns an object of same class as \code{age}
}
\description{
Double normal ogive
}
\examples{
\dontrun{
params=FLPar(a1=1,sl=3,sr=500)
wt2len(params,FLQuant(10))
}
}
\seealso{
\code{\link{sigmoid}}, \code{\link{dnormal}}, \code{\link{logistic}}
}

