% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity.R
\docType{methods}
\name{elas}
\alias{elas}
\alias{elas-method}
\title{elas analysis for reference points}
\usage{
elas(params, sel, fn)
}
\arguments{
\item{params}{parameters}

\item{sel}{selection pattern}

\item{fn}{user supplied function}
}
\value{
\code{FLPar} with elas of reference points by input parameters
}
\description{
Estimates elas
}
\examples{
\dontrun{
#bug
}
}
\seealso{
\code{\link{lhPar}},  \code{\link{lhEql}},  \code{\link{lhRef}}
}
