\name{FLife-package}
\alias{FLife-package}
\alias{FLife}
\docType{package}
\title{
\packageTitle{FLife}
}
\description{
\packageDescription{FLife}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FLife}
\packageIndices{FLife}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FLife}

Maintainer: \packageMaintainer{FLife}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
