% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gascuel.R
\docType{methods}
\name{gascuel}
\alias{gascuel}
\alias{gascuel,FLPar,FLPar-method}
\alias{gascuel,FLPar,FLQuant-method}
\alias{gascuel,FLPar,missing-method}
\alias{gascuel,numeric,FLQuant-method}
\alias{gascuel,numeric,numeric-method}
\alias{gascuel-method}
\title{gascuel}
\usage{
gascuel(params, age, ...)
}
\arguments{
\item{params}{\code{FLPar}}

\item{age}{FLQuant, FLPar or numeric with ages}

\item{...}{any other arguments}
}
\value{
Depends on the value of \code{data}

#'
}
\description{
Gascuel growth equation
}
\details{
Gascuel D., Fonteneau, A., and Capisano, C. (1992).
Modelisation d'une croissance en deux stances chez  #l'albacore (Thunnus albacares) de l'Atlantique Est.
Aquat. Living Resour. 5: 155-172.
}
\examples{
\dontrun{
params=FLPar(gas.a=37.8,gas.b=8.93,gas.c=137.0,
             gas.d=8.93,gas.e=0.808,gas.f=7.49)
len=gascuel(params,age)
}
}
\seealso{
\code{\link{gompertz}}
}

