% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\docType{methods}
\name{gompertz,FLQuant,FLPar-method}
\alias{g}
\alias{gompertz}
\alias{gompertz,FLPar,FLPar-method}
\alias{gompertz,FLQuant,FLPar-method}
\alias{gompertz,FLQuant,numeric-method}
\alias{gompertz,missing,FLPar-method}
\alias{gompertz,numeric,numeric-method}
\alias{gompertz-method}
\title{{gompertz growth equation}}
\usage{
\S4method{gompertz}{FLQuant,FLPar}(age, params, ...)
}
\arguments{
\item{age}{FLQuant, FLPar or numeric with ages}

\item{params}{\code{FLPar} with parameters for \code{linf, a, b}}

\item{...}{any other arguments}
}
\value{
Depends on the value of \code{data}
}
\description{
gompertz growth equation
}
\examples{
\dontrun{
par=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=gompertz(par,age)
}

}

