% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leslie.R
\docType{methods}
\name{r}
\alias{r}
\alias{r,FLPar-method}
\alias{r-method}
\title{lambda}
\usage{
r(object, ...)
}
\arguments{
\item{object}{\code{FLBRP}}

\item{...}{any other arguments}
}
\value{
\code{FLPar}

#'
}
\description{
Estimates population growth rate for a Leslie matrix
}
\examples{
\dontrun{
library(ggplot2)
library(FLBRP)
library(FLife)
data(pars)
pms=lhSim(pars[[1]])
eql=lh(pms)
lsl=leslie(eql)
lambda(lsl)
}
}
\seealso{
\code{\link{leslie}}
}

