% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/len2wt.R
\docType{methods}
\name{len2wt}
\alias{len2wt}
\title{len2wt}
\usage{
len2wt(params, data, ...)
}
\arguments{
\item{params}{}

\item{length}{age FLQuant, FLPar or numeric}
}
\value{
Depends on the value of \code{data}

#'
}
\description{
converts length to weight
}
\examples{
\dontrun{
params=FLPar(a=1,b=3)
len2wt(params,FLQuant(10))
}
}
\seealso{
\code{\code{\link{wt2len}}}
}

