% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lengths}
\alias{lengths}
\title{lengths}
\format{A data frame with 139 rows (i.e. species) and 13 variables:
\describe{
  \item{length}{}
  \item{year}{}
  \item{data}{}
}}
\source{
\url{http://iccat.int/en/accesingdb.htm}
}
\usage{
lengths
}
\description{
A dataset containing lengths by year
}
\examples{
\dontrun{
data(lengths)
head(lengths)}
}
\seealso{
\code{\link{teleosts}}
}
\keyword{datasets}

