% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leslie.R
\docType{methods}
\name{leslie}
\alias{leslie}
\title{leslie}
\usage{
leslie(object, ...)
}
\arguments{
\item{object}{\code{FLBRP}}
}
\value{
\code{matrix}

#'
}
\description{
Creates a Leslie Matrix
}
\examples{
\dontrun{
pms=data(pars[[1]])
eql=gislasim(pms)
lsl=leslie(eql)
}
}
\seealso{
\code{\code{\link{lh}}}
}

