% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lhEql.R
\docType{methods}
\name{lhEql}
\alias{lhEql}
\alias{lhEql,FLPar-method}
\alias{lhEql-method}
\title{lhEql}
\usage{
lhEql(params, ...)
}
\arguments{
\item{params}{an \code{FLPar} object with life history parameters}

\item{...}{any other arguments that include,
growth function for growth,
m function for natutal mortality,
mat function for proportion mature-at-age,
sel function for selectivity-at-age,
sr character for stock recruitment relationship,
range age range,
spwn proportion of year when spawning occurrs, i.e. level of natural mortality prior to spawning,
fish proportion of year when fishing happens,
units units for FLQuant slots}
}
\value{
\code{FLBRP}
}
\description{
Takes an \code{FLPar} object with life history and selectivity parameters
and generates an corresponding \code{FLBRP} object. Can uses a range of functional forms
}
\examples{
\dontrun{
par=lhSim(FLPar(linf=100))
}
}
\seealso{
\code{\link{vonB}} \code{\link{lorenzen}} \code{\link{sigmoid}}
}

