% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLife-lhPar.R
\docType{methods}
\name{lhPar}
\alias{lhPar}
\title{Generates life history parameters}
\usage{
lhPar(..., m = list(model = "gislason", params = c(m1 = 0.55, m2 = -1.61,
  m3 = 1.44)))
}
\arguments{
\item{params}{\code{FLPar} object with parameters for life history equations and selection pattern.
Need Linfinity to estimate other parameters, if any other parameters supplied in \code{code} then
these are not provided by the algorithm}

\item{t0}{of von Bertalanffy. This is a default that isnt normally derived from life history theory, as are the following args.}

\item{a}{coefficient of length weight relationship}

\item{b}{exponent of length weight relationship}

\item{ato95}{age at which 95\% of fish are mature, offset to age at which 50\% are mature}

\item{s}{steepness of stock recruitment relationship}

\item{v}{virgin biomass}

\item{sel1}{selectivity-at-age parameter for double normal, age at maximum selectivity by default set to same as age at 100\% mature}

\item{sel2}{selectivity-at-age parameter for double normal, standard deviation of lefthand limb of double normal, by default 5}

\item{sel3}{selectivity-at-age parameter for double normal, standard deviation of righthand limb of double normal, by default 5000}

\item{sl}{obsolete now replaced by sel2}

\item{sr}{obsolete now replaced by sel3}

\item{m1}{m-at-age parameter by default for Gislason empirical relationship}

\item{m2}{m-at-age parameter, by default for Gislason empirical relationship}

\item{m3}{m-at-age parameter, by default for Gislason empirical relationship}
}
\value{
object of class \code{FLPar} with missing parameters calculated from life history theory
}
\description{
Uses life history theory to derive parameters for biological relationships, i.e. or growth, 
maturity, natural mortality. Selectivity by default is set so age at peak selectivity is the 
same as age at 50/% mature (a50) As a minimum all `lhPar` requires is `linf` the asymptotic 
length of the von Bertalannfy growth equation.
}
\examples{
\dontrun{
#COMPARE with output of FLife::lhPar

x <- as(lhpar(linf=100), 'list')
x <- x[sort(names(x))]
y <- as(lhPar(FLPar(linf=100)), 'list')
y <- y[sort(names(y))]

all.equal(x,y)

for(i in seq(length(x)))
   cat(names(x[i]), ":", unlist(x[i]), "-", names(y[i]), ":", unlist(y[i]), "\\n")
   
 # CALL with iters
 lhpar(FLPar(linf=100), v=rnorm(100, 300, 200))
 
 lhPar(FLPar(linf=rnorm(100, 80, 10)))
 lhPar(FLPar(linf=100, v=rnorm(100, 300, 200)))
 lhPar(FLPar(linf=100), FLPar(v=rnorm(100, 300, 200)))
 lhPar(FLPar(linf=100, v=rnorm(100, 300, 200)), t0=-1, data.frame(a=1,b=7))
 
 attributes(lhpar(FLPar(linf=100), v=rnorm(100, 300, 200)))$mmodel
}

}
\seealso{
\code{\link{loptAge}}, \code{\link{lhRef}}, \code{\link{lhPar}}, \code{\link{lhEql}}
}
