% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhRef.R
\docType{methods}
\name{lhRef}
\alias{lhRef}
\title{Reference points based on life histories}
\usage{
lhRef(params, m = function(length, params) exp(0.55) * (length^-1.61) \%*\%
  (params["linf"]^1.44) \%*\% params["k"], sr = "bevholt", range = c(min =
  0, max = 40, minfbar = 1, maxfbar = 40, plusgroup = 40), what = c("r",
  "lopt", "rc", "sk", "spr0"))
}
\arguments{
\item{params}{\code{FLPar}}

\item{m}{function for natural mortality}

\item{sr}{\code{character} with stock recruitment relationship, e.g. "bevholt","ricker",...}

\item{range}{ages used i.e. c(min=0,max=40,minfbar=1,maxfbar=40,plusgroup=40)}

\item{what}{quantities to calculate "r","lopt","rc","sk","spr0"}
}
\value{
object of class \code{FLPar} with reference points, i.e r, rc, sk, lopt,
}
\description{
\code{lhRef} calculates a variety of reference points i.e.  population growth rate at small 
population sizs (r), and at Bmsy (rc), ratio of virgin biomass to Bmsy (sk), life 
time reproductive output (srp0) and reproductive output at $B_{MSY}$ ($srp_{msy}$)
}
\examples{
\dontrun{
params=FLPar(linf=100,t0=0,k=.4)
params=lhPar(params)
lhRef(params)
}
}
\seealso{
\code{\link{lhPar}}, \code{\link{lhEql}}
}

