% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logistic.R
\docType{methods}
\name{logistic}
\alias{logistic}
\title{logistic}
\usage{
logistic(age, params, ...)
}
\arguments{
\item{age}{}

\item{par}{}
}
\description{
logistic function
}
\examples{
\dontrun{
par=lhPar(FLPar(linf=100))
age=FLQuant(1:10,dimnames=list(age=1:10))
mat=logistic(par,age)
}
}
\seealso{
\code{\link{gompertz}}
}

