% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lopt.R
\docType{methods}
\name{lopt}
\alias{lopt}
\title{lopt}
\usage{
lopt(param, ...)
}
\arguments{
\item{param}{}

\item{fnM}{natural mortality}
}
\value{
numeric with length

#'
}
\description{
Finds length at maximum biomass
}
\examples{
\dontrun{
data(pars)
lopt(pars[[1]])
}
}
\seealso{
\code{\code{\link{vonB}}}
}

