% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt.R
\docType{methods}
\name{loptAge,FLPar-method}
\alias{loptAge}
\alias{loptAge,FLPar-method}
\alias{loptAge-method}
\title{loptAge}
\usage{
\S4method{loptAge}{FLPar}(params, m = function(length, params) params["m1"]
  \%*\% (exp(log(length) \%*\% params["m2"])), growth = vonB, ...)
}
\arguments{
\item{params}{FLPar}

\item{m}{A function, i.e. gislason}

\item{growth}{A function, i.e. vonB}

\item{...}{any other arguments}
}
\value{
FLPar with length at maximum biomass
}
\description{
Finds length at maximum biomass
}
\details{
There are several ways to calculate \deqn{L_{opt}}, i.e.
i) \deqn{{2/3}^{rds}  L_{\infty}}
ii) \deqn{L_{\infty}\frac{3}{3+k/m}}
iii) by maximising the biomass of
iv) from an FLBRP object by fishing at F=0 and finding age where biomass is a maximum
}
\examples{
\dontrun{
data(pars)
lopt(pars[[1]])
}
}
\seealso{
\code{\link{vonB}}
}

