% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mat-dd.R
\docType{methods}
\name{matdd}
\alias{matdd}
\alias{matdd,FLQuant,FLPar-method}
\title{matdd}
\usage{
matdd(age, params, ...)
}
\arguments{
\item{age}{ages}

\item{params}{an \code{FLPar} with two values; i.e. a equal to M at unit mass and b a power term; defaults are a=0.3 and b=-0.288}

\item{...}{other arguments, such as scale, e.g. stock numbers now relative to a reference level, e.g. at virgin biomass and k steepness of relationship}
}
\description{
Logistic ogive for proportion mature-at-age, modified to explicitly included maturity
as a function of numbers in a cohort, i.e. density dependence
}
\details{
The relationship can be explained by population density, since as fish grow they
also die and so there is potentially less competition for resources between larger and
older fish. Density dependence can be modelled by a logistic function, a sigmoid
curve (or S shaped) curve, with equation

f(x)=L/(1+exp(-k(x-x0)))

where
e  is the natural logarithm base (also known as Euler's number),
x0 is the x-value of the sigmoid's midpoint,
L  is the curve's maximum value, and
k  the steepness of the curve.

Combining the two functions gives

   O=aL/(1+exp(-k(n-ref)))*wt^b;
}
\examples{
\dontrun{
library(FLBRP)
library(FLife)

data(ple4)

sr=fmle(as.FLSR(ple4,model="bevholtSV"), fixed=list(s=.75,spr0=spr0(FLBRP(ple4))))
#control=list("silent"=TRUE))

eql=brp(FLBRP(ple4,sr=ab(sr)))
fbar(eql)=FLQuant(c(seq(0,4,length.out=101)*refpts(eql)["msy","harvest"]))
stk=as(eql,"FLStock")

fbar(eql)=fbar(eql)[,1]
scale=(stock.n(stk)\%-\%stock.n(eql))\%/\%stock.n(eql)
par=lhPar(FLPar(linf=100))

scale=exp(noise(1,m(stk),sd=.2,b=0.75))

ggplot(as.data.frame(mat,drop=TRUE))+
   geom_line(aes(age,data,group=year,col=factor(year)))+
   theme(legend.position="none")

 }
}
\seealso{
\code{\link{logistic},\link{mdd}}
}

