% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\docType{methods}
\name{moment,numeric-method}
\alias{moment}
\alias{moment,numeric-method}
\alias{moment-method}
\title{moment}
\usage{
\S4method{moment}{numeric}(object, n = rep(1, length(object)), na.rm = T)
}
\arguments{
\item{object}{a vector holding a time series}

\item{n}{number of observations equal to length og object by default}

\item{na.rm}{\code{boolean} whether to remove NAs, TRUE by default}

\item{...}{any other arguments, i.e. x,n=rep(1,length(x)),na.rm=T}
}
\value{
a \code{vector} with the inter-annual variation each time step
}
\description{
aaa
}
\examples{
x=1
}

