% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/noise.R
\docType{methods}
\name{noise}
\alias{noise}
\title{noise}
\usage{
noise(n, len, ...)
}
\arguments{
\item{n}{number of iterations}

\item{len}{an \code{FLQuant}}

\item{sd}{standard deviation}

\item{b}{autocorrelation parameter, a real number in [0,1]}

\item{trunc}{get rid of first values equal to trunc, i.e. to allow burn in}
}
\value{
A vector with autocorrelation equal to B.
}
\description{
A noise generator
}
\examples{
\dontrun{
white <- noise(1000,sd=.3,b=0)
plot(white)
acf(white)

red <- noise(1000,sd=.3,b=0.8)
plot(red)
acf(white)

blue <- noise(1000,sd=.3,b=-0.8)
plot(blue)
acf(blue)
}
}
\references{
Ranta and Kaitala 2001 Proc. R. Soc.
vt = b * vt-1 + s * sqrt(1 - b^2)
s is normally distributed random variable with mean = 0
b is the autocorrelation parameter
}

