% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omOut.R
\docType{methods}
\name{omSmry}
\alias{omSmry}
\alias{omSmry-method}
\alias{omSmry,FLStock-method}
\title{omSmry}
\usage{
omSmry(object, ...)
}
\arguments{
\item{object}{\code{FLStock}}

\item{...}{any other arguments}
}
\description{
create time series of summary statistics from \code{FLStock} relative to reference points
}
\details{
omSmry
}
\examples{
\dontrun{
res=omSmry(om,eq)
}
}
\seealso{
smryStat
}
\author{
Laurence Kell, Sea++
}
