% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/popPar.R
\docType{methods}
\name{popPar}
\alias{popPar}
\title{popPar}
\usage{
popPar(params, m = function(length, params) params["m1"] \%*\%
  (exp(log(length) \%*\% params["m2"])))
}
\arguments{
\item{params}{\code{FLPar}}

\item{age}{FLQuant, FLPar or numeric with ages}

\item{length}{FLQuant, FLPar or numeric with length, if supplied as a named paramreter
instead  of age then calculates ages.}
}
\value{
Depends on the value of \code{data}

#'
}
\description{
Von Bertalanffy growth equation
}
\examples{
\dontrun{
params=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=vonB(age,params)
age=vonB(params,length=len)
}
}
\seealso{
\code{\link{gascuel}}
}

