% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/powh.R
\docType{methods}
\name{powh}
\alias{pow-h}
\alias{powh}
\alias{unbin}
\title{powh}
\usage{
powh(len, n, weights = TRUE, fromMode = FALSE)

unbin(x)
}
\arguments{
\item{len}{lengths}

\item{n}{frequqncies}

\item{weights=TRUE}{}

\item{fromMode=FALSE}{}

\item{len}{vector with length distribution}

\item{n}{vector with numbers in each length bin}

\item{x;}{a vector of with intervals as names}
}
\value{
a \code{data.frame} \code{mn} (mean), \code{diff} (difference),
\code{len} (length) and \code{n} (frequency)

a \code{data.frame} with left and right boundaries and mid points.
}
\description{
Estimates growth and mortality parameters from length frequency data.

For a vector with labels corresponding to intervals i.e. \code{"(0,10]"}
 returns a data.frame with left and right boundaries and mid point.
}
\details{
Beverton and Holt (1956) developed a method to estimate population parameters
 such as total mortality (Z) from length data i.e.

\deqn{Z=K\frac{L_{infinity}-\overline{L}}{\overline{L}-L'}}

Powell (1979) then developed a method, extended by Wetherall et al. (1987),
to estimate growth and mortality parameters. This assumes that the right hand tail
of a length frequency distribution is determined by the asymptotic length L
and the ratio between Z and the growth rate K.

The Beverton and Holt methods assumes good estimates for K and $L_{infinity}$,
while the Powell-Wetherall method only requires an estimate
of K, since $L_{infinity}$ is estimated by the method as well as Z/K. These method
therefore provide estimates for Z/K, if K is unknown and
Z if K is known.

As well as assuming that growth follows the von Bertalanffy growth function,
it is also assumed that the population is in a steady state
with constant exponential mortality, no changes in selection pattern of the
fishery and constant recruitment.

In the Powell-Wetherall method L' can take any value between the
smallest and largest sizes. Equation 1 then provides a series of estimates of
Z and since

\deqn{
\overline{L}-L'=a+bL'
}

a and b can be estimated by a regression analysis where

\deqn{b={-K}/{Z+K}}
\deqn{a=-bL_{infinity}}

Therefore plotting \deqn{\overline{L}-L} against $L'$ provides an estimate
of $L_{infinity}$ and Z/K from

\deqn{L_{infinity}=-a/b}
\deqn{Z/K={-1-b}/{b}}

If K is known Z can also be esimated
}
\examples{
\dontrun{
data(bonLn)
rslt=with(subset(bonLn,year==2013), powh(len,n))
}
x=summary(cut(runif(100),seq(0,1,.1)))
unbin(x)
}
\references{
R. Beverton and S. Holt. Review of method for estimating mortality rates in
 exploited fish populations, with special reference to sources of bias in
 catch sampling. Rapports et Proces-Verbaux., 140(1): 67--83, 1956.

 D. G. Powell.
 Estimation of mortality and growth parameters from the length
 frequency of a catch [model].
 \emph{Rapports et Proces-Verbaux des Reunions}, 175, 1979.

 J. Wetherall, J. Polovina, and S. Ralston.
 Estimating growth and mortality in steady-state fish stocks from
 length-frequency data.
 \emph{ICLARM Conf. Proc}, pages 53--74, 1987.
}

