% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refs.R
\name{refs}
\alias{refs}
\title{refs}
\usage{
refs(object, ...)
}
\arguments{
\item{object}{\code{FLStock}}
}
\value{
\code{data.frame} 
'
}
\description{

}
\examples{
\dontrun{

 library(FLCore)
 library(FLBRP)
 library(ggplotFL)
 library(FLife)
 library(popbio)
 
 data(ple4)
 
 refs(ple4)
} 
}

