% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richards.R
\docType{methods}
\name{richards,FLQuant,FLPar-method}
\alias{richards}
\alias{richards,FLPar,FLPar-method}
\alias{richards,FLQuant,FLPar-method}
\alias{richards,FLQuant,numeric-method}
\alias{richards,missing,FLPar-method}
\alias{richards,numeric,numeric-method}
\alias{richards-method}
\title{von Bertalanffy growth curve}
\usage{
\S4method{richards}{FLQuant,FLPar}(age, params, ...)
}
\arguments{
\item{age}{FLQuant, FLPar or numeric object with values corresponding to ages}

\item{params}{\code{FLPar} object with parameters \code{linf}, \code{k} and \code{t0}}

\item{...}{other arguments}
}
\value{
Returns an object of same class as \code{age}  e.g. \code{FLQuant}
}
\description{
Von Bertalanffy growth equation
}
\examples{
\dontrun{
params=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=richards(age,params)

#inverse growth curve
richards(params=params,length=len)
}
}
\seealso{
\code{\link{vonB}}, \code{\link{gompertz}}, \code{\link{gascuel}}
}

