% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\docType{methods}
\name{rod,FLQuant-method}
\alias{rod}
\alias{rod,FLQuant-method}
\alias{rod-method}
\title{rod}
\usage{
\S4method{rod}{FLQuant}(object, ...)
}
\arguments{
\item{object}{an object of class \code{FLQuant}}

\item{...}{any other arguments}
}
\description{
Regime shifts
Evidence for regime shifts are explored using a a sequential t-test algorithm 
(STARS; \cite{rodionov2004sequential}) as modified by Szuwalski et al., (submitted)
}
\details{
Returns a data.frame
}
\examples{
\dontrun{
 object=rlnorm(1,FLQuant(0,dimnames=list(year=1:30)),.3)
 pg=rod(object) 
 ggplot(object) +
     geom_polygon(aes(year,data,group=regime),
         fill="lavender",col="blue",
         lwd=.25,data=pg,alpha=.75)+
     geom_point(aes(year,data))+
     geom_line(aes(year,data))
}

}

