% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sigmoid.R
\docType{methods}
\name{sigmoid}
\alias{sigmoid}
\title{sigmoid}
\usage{
sigmoid(age, param, ...)
}
\arguments{
\item{age}{FLQuant, FLPar or numeric with ages}

\item{param}{}
}
\value{
Depends on the type of \code{age}
}
\description{
sigmoid ogive
}
\details{
The sigmoid ogive is an S-shaped or sigmoid curve or sigmoidal functions,
Verhulst hypothesizes that small populations increase geometrically, because the supply of resources exceeds demand. Then, as supply and demand balance, population growth is constant. Finally, as demand exceeds supply, population growth decreases at the same rate that it had increased. Verhulst describes this process with an equation that enables him to predict when a population will reach any given size (see Verhulst's Figure):
}
\examples{
\dontrun{
param=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=sigmoid(age,param)
age=sigmoid(param,length=len)
}
}
\seealso{
\code{\code{\link{dnormal},\link{knife}}}
}

