% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lh-srr-sv.R
\docType{methods}
\name{sv,FLPar,character-method}
\alias{sv,FLPar,character-method}
\title{Calculates steepness and virgin biomass}
\usage{
\S4method{sv}{FLPar,character}(x, model, spr0 = NA)
}
\arguments{
\item{x}{\code{FLPar} with a and b}

\item{model}{\code{character} with name of stock recruitment relationship, by default "bevholt"}

\item{spr0}{\code{} spawner per recruit at F=0}

\item{...}{any other arguments}
}
\value{
\code{FLPar} with values for steepness (s) and virgin biomass (v)
}
\description{
Calculates steepness and virgin biomass given a and b for a Beverton and Holt SRR
}
\examples{
\dontrun{
#bug
params=FLPar(a=37.8,b=8.93)
sv(params,"bevholt",.4)
}
}
