% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teleost}
\alias{teleost}
\title{teleost}
\format{A data frame with 139 rows (i.e. species) and 13 variables:
\describe{
  \item{species,genus,family,order,class}{} 
  \item{linf}{asymptotic length of the Von Bertalanffy growth equation}    
  \item{k}{rate at which length reaches linf}       
  \item{t0}{adjusts the growth equation for the initial size at birth}      
  \item{l50}{length at which 50\% are mature}     
  \item{a}{scaling factor of the length weight relationship}       
  \item{b}{exponent of the length weight relationship}       
  \item{temp}{temperature}    
  \item{habit}{e.g. pelagic or demeresal}   
}}
\source{
\url{http://www.fishbase.org/}
}
\usage{
teleost
}
\description{
A dataset containing life history parameters for a range of teleost species
}
\examples{
\dontrun{
data(teleost)
head(teleost)} 
}
\seealso{
\code{\link{lengths}}
}
\keyword{datasets}

