% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powh.R
\docType{methods}
\name{unbin}
\alias{unbin}
\title{unbin frequency distributions}
\usage{
unbin(x)
}
\arguments{
\item{x}{a vector of with intervals as names}
}
\value{
a \code{data.frame} with left and right boundaries and mid points.
}
\description{
For a vector with labels corresponding to intervals i.e. \code{"(0,10]"}
 returns a data.frame with left and right boundaries and mid point.
}
\examples{
x=summary(cut(runif(100),seq(0,1,.1)))
unbin(x)
}

