% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vonB.R
\docType{methods}
\name{vonB}
\alias{vonB}
\alias{vonB,FLPar,FLPar-method}
\alias{vonB,FLQuant,FLPar-method}
\alias{vonB,FLQuant,numeric-method}
\alias{vonB,missing,FLPar-method}
\alias{vonB,numeric,numeric-method}
\title{vonB}
\usage{
vonB(age, params, ...)
}
\arguments{
\item{age}{FLQuant, FLPar or numeric with ages}

\item{params}{\code{FLPar}}

\item{...}{any other arguments, e.g. length FLQuant, FLPar or numeric with length, if supplied as a named paramreter
instead  of age then calculates ages.}
}
\value{
Depends on the value of \code{data}

#'
}
\description{
Von Bertalanffy growth equation
}
\examples{
\dontrun{
params=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=vonB(age,params)
age=vonB(params,length=len)
}
}
\seealso{
\code{\link{gascuel}}
}

