% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wklife}
\alias{wklife}
\title{wklife}
\format{
A data frame with 15 rows (i.e. species) and 13 variables:
\describe{
  \item{species}{} 
  \item{name}{} 
  \item{area}{} 
  \item{stock}{} 
  \item{sex}{} 
  \item{a}{scaling factor of the length weight relationship}       
  \item{b}{exponent of the length weight relationship}       
  \item{linf}{asymptotic length of the Von Bertalanffy growth equation}    
  \item{k}{rate at which length reaches linf}       
  \item{t0}{adjusts the growth equation for the initial size at birth}      
  \item{lmax}{asymptotic length of the Von Bertalanffy growth equation}    
  \item{l50}{length at which 50\% are mature}     
  \item{a50}{age at which 50\% are mature}     
}
}
\description{
A dataset containing life history parameters for a range of teleost species
}
\seealso{
\code{\link{teleost}}
}
