% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wt2len.R
\docType{methods}
\name{wt2len}
\alias{wt2len}
\title{wt2len}
\usage{
wt2len(wt, param, ...)
}
\arguments{
\item{wt}{FLQuant, FLPar or numeric with length}

\item{param}{FLPar with a and b, i.e. condition and scaling factors}
}
\value{
Depends on the value of \code{wt}
}
\description{
converts weight to length
}
\examples{
\dontrun{
param=FLPar(a=1,b=3)
wt2len(FLQuant(10),param)
}

if (!isGeneric("rgamma"))
}
\seealso{
\code{\link{len2wt}}
}

