% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ages.R
\docType{methods}
\name{ages,FLQuant-method}
\alias{ages}
\alias{ages,FLCohort-method}
\alias{ages,FLQuant-method}
\alias{ages-method}
\title{ages}
\usage{
\S4method{ages}{FLQuant}(object, season = NULL)
}
\arguments{
\item{object}{\code{FLQuant} or \code{FLCohort}}

\item{season}{the season dim}

\item{...}{any other arguments}
}
\value{
\code{FLQuant} or \code{FLCohort}
}
\description{
Creates FLQuant/FLCohort with ages as entries
}
\examples{
\dontrun{
data(ple4)
ages(m(ple4))}
}
\seealso{
\code{\link{knife}} \code{\link{gascuel}} \code{\link{sigmoid}} \code{\link{gompertz}} \code{\link{vonB}} \code{\link{dnormal}} \code{\link{logistic}}
}

