% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cas}
\alias{cas}
\title{cas}
\format{A data frame with catch-at-length data:
\describe{
  \item{year}{ of capture}
  \item{len}{ at capture}   
  \item{n}{ frequency}   
}}
\source{
\url{http://iccat.int/en/accesingdb.htm}
}
\usage{
cas
}
\description{
A dataset containing lengths by year
}
\examples{
\dontrun{
data(cas)
head(cas)} 
}
\seealso{
\code{\link{teleost}}
}
\keyword{datasets}

