% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\docType{methods}
\name{cc,numeric,numeric-method}
\alias{cc}
\alias{cc,FLQuant,missing-method}
\alias{cc,missing,FLQuant-method}
\alias{cc,numeric,numeric-method}
\alias{cc-method}
\title{cc}
\usage{
\S4method{cc}{numeric,numeric}(age, n, ...)
}
\arguments{
\item{age}{age}

\item{n}{frequency}

\item{...}{any other arguments}
}
\value{
Depends on the value of \code{data}
}
\description{
Catch curve analysis
}
\examples{
\dontrun{
params=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=vonB(params,age)
age=vonB(params,length=len)
}
}
\seealso{
\code{\link{powh}}
}

