% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\docType{methods}
\name{cc,numeric,numeric-method}
\alias{cc,numeric,numeric-method}
\alias{cc}
\alias{cc-method}
\alias{cc,missing,FLQuant-method}
\alias{cc,FLQuant,missing-method}
\title{Catch curve analysis}
\usage{
\S4method{cc}{numeric,numeric}(age, n, ...)
}
\arguments{
\item{age}{age}

\item{n}{frequency}

\item{...}{any other arguments}
}
\value{
Returns an object of same class \code{age}
}
\description{
Catch curve analysis
}
\examples{
\dontrun{
data(ple4)
ctc=as.data.frame(catch.n(ple4))
dat=cc(age=ctc$age,n=ctc$data)
head(dat)
}
}
\seealso{
\code{\link{powh}}

\code{\link{powh}}
}
