% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cc.R
\docType{methods}
\name{cc}
\alias{cc}
\title{cc}
\usage{
cc(age, n, ...)
}
\arguments{
\item{age}{}

\item{n}{}
}
\value{
Depends on the value of \code{data}
}
\description{
Catch curve analysis
}
\examples{
\dontrun{
params=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=vonB(params,age)
age=vonB(params,length=len)
}
}
\seealso{
\code{\code{\link{powh}}}
}

