\name{data}
\alias{eql}
\alias{ln}
\alias{nln}
\alias{pars}
\title{FLife datasets}

\description{
  \itemize{
    \item{eql}{a}
    \item{ln}{a}
    \item{nlnl}{a}
    \item{pars}{a}
    
  Datasets can be loaded by issuing the \code{data} command, as in \code{data(ln)}.
  }
  }
    
\seealso{
  \linkS4class{FLStock}, \linkS4class{FLSR}, \linkS4class{FLBiol}
  }

\references{nnnnn}

\keyword{datasets}
