% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity.R
\docType{methods}
\name{elasticity}
\alias{elasticity}
\alias{elasticity-method}
\title{Elasticity analysis for reference points}
\usage{
elasticity(params, sel, fn)
}
\arguments{
\item{params}{parameters}

\item{sel}{selection pattern}

\item{fn}{user supplied function}
}
\value{
\code{FLPar} with elasticity of reference points by input parameters
}
\description{
Estimates elasticity
}
\examples{
\dontrun{
#bug
}
}
\seealso{
\code{\link{lhPar}},  \code{\link{lhEql}},  \code{\link{lhRef}}
}
