% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/elasticity.R
\docType{methods}
\name{elasticity}
\alias{elasticity}
\alias{elasticity-method}
\title{elasticity}
\usage{
elasticity(par, sel, fn)
}
\arguments{
\item{par}{}

\item{sel}{}

\item{fn}{}
}
\value{
\code{FLPar}

#'
}
\description{
Estimates elasticity
}
\examples{
\dontrun{
FLPar(linf=200)
}
}
\seealso{
\code{\link{lhPar}}
}

