% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/elasticity.R
\docType{methods}
\name{elasticity}
\alias{elasticity}
\alias{elasticity-method}
\title{elasticity}
\usage{
elasticity(params, sel, fn)
}
\arguments{
\item{params}{parameters}

\item{sel}{selection pattern}

\item{fn}{user supplied function}
}
\value{
\code{FLPar}

#'
}
\description{
Estimates elasticity
}
\examples{
\dontrun{
library(ggplot2)
library(FLCore)
library(FLBRP)
library(FLife)
data(pars)
pms=lhSim(pars[[1]])
eql=lh(pms,range = c(min=0,max=8, minfbar=1,maxfbar=8,plusgroup=8))
lsl=leslie(eql,fbar=c(refpts(eql)["crash","harvest"]))
}
}
\seealso{
\code{\link{lhPar}}
}

