% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gascuel.R
\docType{methods}
\name{gascuel,FLQuant,FLPar-method}
\alias{gascuel}
\alias{gascuel,FLPar,FLPar-method}
\alias{gascuel,FLPar,FLQuant-method}
\alias{gascuel,FLPar,missing-method}
\alias{gascuel,FLQuant,FLPar-method}
\alias{gascuel,FLQuant,missing-method}
\alias{gascuel,missing,FLPar-method}
\alias{gascuel,missing,missing-method}
\alias{gascuel,numeric,FLPar-method}
\alias{gascuel,numeric,FLQuant-method}
\alias{gascuel,numeric,missing-method}
\alias{gascuel,numeric,numeric-method}
\alias{gascuel-method}
\title{Gascuel growth curve}
\usage{
\S4method{gascuel}{FLQuant,FLPar}(age, params, ...)
}
\arguments{
\item{age}{FLQuant, FLPar or numeric with ages}

\item{params}{\code{FLPar}}

\item{...}{any other arguments}
}
\value{
Returns a class of same type as \code{age} e.g. \code{FLQuant}
}
\description{
Gascuel growth equation
}
\details{
Gascuel D., Fonteneau, A., and Capisano, C. (1992).
Modelisation d'une croissance en deux stances chez  #l'albacore (Thunnus albacares) de l'Atlantique Est. 
Aquat. Living Resour. 5: 155-172.
}
\examples{
\dontrun{
gascuel(10)
}
}
\seealso{
\code{\link{gompertz}}, \code{\link{vonB}}
}

