% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLife-m.R
\docType{methods}
\name{gislason,FLQuant,numeric-method}
\alias{gislason,FLQuant,numeric-method}
\alias{gislason}
\alias{gislason-method}
\alias{gislason,FLQuant,FLPar-method}
\alias{gislason,FLQuant,missing-method}
\title{gislason}
\usage{
\S4method{gislason}{FLQuant,numeric}(length, params, a = 0.55, b = 1.44, c = -1.61, ...)
}
\arguments{
\item{length}{mass at which M is to be predicted}

\item{params}{\code{FLPar} with two values; i.e. a equal to M at unit mass 
and b a power term; defaults are a=0.3 and b=-0.288}

\item{a}{0.55}

\item{b}{1.44}

\item{c}{-1.61}

\item{...}{any other arguments}
}
\description{
gislason natural mortality relatoinship estimate M as a function of length. 
M=a*length^b;
}
\examples{
\dontrun{
params=lhPar(FLPar(linf=111))
len=FLQuant(c( 1.90, 4.23, 7.47,11.48,16.04,20.96,26.07,31.22,
               36.28,41.17,45.83,50.20,54.27,58.03,61.48,64.62),
             dimnames=list(age=1:16))
gislason(length,params)
}
}
\seealso{
\code{\link{lorenzen}}
}
