% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gislason.R
\docType{methods}
\name{gislason}
\alias{gislason}
\alias{gislason,FLQuant,FLPar-method}
\alias{gislason,FLQuant,missing-method}
\alias{gislason,FLQuant,numeric-method}
\alias{gislason-method}
\title{gislason}
\usage{
gislason(length, params, ...)
}
\arguments{
\item{length}{mass at which M is to be predicted}

\item{params}{\code{FLPar} with two values; i.e. a equal to M at unit mass
and b a power term; defaults are a=0.3 and b=-0.288}

\item{...}{any other arguments}
}
\description{
gislason natural mortality relatoinship estimate M as a function of weight.
M=a*length^b;
}
\examples{
\dontrun{
length=FLQuant(c( 1.90, 4.23, 7.47,11.48,16.04,20.96,26.07,31.22,
               36.28,41.17,45.83,50.20,54.27,58.03,61.48,64.62),
             dimnames=list(age=1:16))
gislason(mass)
}
}
\seealso{
\code{\link{gislason}}
}

