% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gompertz.R
\docType{methods}
\name{gompertz}
\alias{gompertz}
\alias{gompertz,FLPar,FLPar-method}
\alias{gompertz,FLQuant,FLPar-method}
\alias{gompertz,FLQuant,numeric-method}
\alias{gompertz,missing,FLPar-method}
\alias{gompertz,numeric,numeric-method}
\title{gompertz}
\usage{
gompertz(age, params, ...)
}
\arguments{
\item{age}{FLQuant, FLPar or numeric with ages}

\item{params}{\code{FLPar} with parameters for \code{linf, a, b}}

\item{...}{any other arguments}
}
\description{
Gompertz growth equation
}
\examples{
\dontrun{
par=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=gompertz(par,age)
}
}
\seealso{
\code{\link{gompertz}}
}

