% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\docType{methods}
\name{gompertz,FLQuant,FLPar-method}
\alias{g}
\alias{gompertz}
\alias{gompertz,FLPar,FLPar-method}
\alias{gompertz,FLQuant,FLPar-method}
\alias{gompertz,FLQuant,numeric-method}
\alias{gompertz,missing,FLPar-method}
\alias{gompertz,numeric,numeric-method}
\alias{gompertz-method}
\title{Gompertz growth equation}
\usage{
\S4method{gompertz}{FLQuant,FLPar}(age, params, ...)
}
\arguments{
\item{age}{FLQuant, FLPar or numeric with ages}

\item{params}{\code{FLPar} with parameters for \code{linf, a, k}}

\item{...}{any other arguments}
}
\value{
Returns an object of same class as \code{age} e.g. \code{FLQuant}
}
\description{
gompertz growth equation
}
\examples{
\dontrun{
params=FLPar(linf=100,a=2,b=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
gompertz(age,params)
}

}
\seealso{
\code{\link{gascuel}}, \code{\link{vonB}}
}

