% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knife.R
\docType{methods}
\name{knife,FLQuant,FLPar-method}
\alias{knife,FLQuant,FLPar-method}
\alias{knife}
\alias{knife-method}
\alias{knife,FLPar,FLPar-method}
\alias{knife,numeric,numeric-method}
\alias{knife,FLQuant,numeric-method}
\title{knife edge ogive}
\usage{
\S4method{knife}{FLQuant,FLPar}(age, params, ...)
}
\arguments{
\item{age}{FLQuant, FLPar or numeric with ages}

\item{params}{\code{FLPar}}

\item{...}{any other arguments}
}
\value{
returns an object of same type as \code{age} e.g. \code{FLQuant}
}
\description{
A method to simulate a knife edge ogive where at a given age the proportion chnages from 0 to 1
}
\details{
The knife ogive is an S-shaped or knife curve or knifeal functions, 
Verhulst hypothesizes that small populations increase geometrically, because the supply of 
resources exceeds demand. Then, as supply and demand balance, population growth is constant. 
Finally, as demand exceeds supply, population growth decreases at the same rate that it had 
increased. Verhulst describes this process with an equation that enables him to predict when 
a population will reach any given size (see Verhulst's Figure):
}
\examples{
\dontrun{
params=FLPar(a1=4)
age=FLQuant(1:10,dimnames=list(age=1:10))
knife(age,params)
}
}
\seealso{
\code{\link{sigmoid}},  \code{\link{dnormal}}, \code{\link{logistic}}
}
