% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knife.R
\docType{methods}
\name{knife}
\alias{knife}
\title{knife}
\usage{
knife(age, param, ...)
}
\arguments{
\item{age}{FLQuant, FLPar or numeric with ages}

\item{param}{}

\item{length}{FLQuant, FLPar or numeric with length, if supplied as a named paramreter
instead  of age then calculates ages.}
}
\value{
Depends on the value of \code{data}
}
\description{
knife edge ogive
}
\details{
The knife ogive is an S-shaped or knife curve or knifeal functions,
Verhulst hypothesizes that small populations increase geometrically, because the supply of resources exceeds demand. Then, as supply and demand balance, population growth is constant. Finally, as demand exceeds supply, population growth decreases at the same rate that it had increased. Verhulst describes this process with an equation that enables him to predict when a population will reach any given size (see Verhulst's Figure):
}
\examples{
\dontrun{
param=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=knife(age,param)
age=knife(param,length=len)
}
}
\seealso{
\code{\link{dnormal},\link{sigmoid}}
}

