% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leslie.R
\docType{methods}
\name{r,FLPar,missing-method}
\alias{r,FLPar,missing-method}
\alias{r-method}
\alias{r,FLPar-method}
\title{Population growth rate}
\usage{
\S4method{r}{FLPar,missing}(m, fec, ...)
}
\arguments{
\item{m}{\code{FLPar}}

\item{fec}{\code{missing}}

\item{...}{any other arguments}
}
\value{
\code{FLPar} with growth rate a small population size
}
\description{
Estimates population growth rate for a Leslie matrix
}
\examples{
\dontrun{
library(popbio)
eql=lhEql(lhPar(FLPar(linf=100)))
L=leslie(eql)
lambda(L[drop=TRUE])
}

}
\seealso{
\code{\link{leslie}}, \code{\link{lhRef}}
}
