% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leslie.R
\docType{methods}
\name{r,FLPar,missing-method}
\alias{r,FLPar,missing-method}
\alias{r,FLPar-method}
\alias{r-method}
\title{r}
\usage{
\S4method{r}{FLPar,missing}(m, fec, ...)
}
\arguments{
\item{m}{\code{FLPar}}

\item{fec}{\code{missing}}

\item{...}{any other arguments}
}
\value{
\code{FLPar}
}
\description{
Estimates population growth rate for a Leslie matrix
}
\examples{
\dontrun{
library(ggplot2)
library(FLBRP)
library(FLife)
data(pars)
pms=lhSim(pars[[1]])
eql=lh(pms)
lsl=leslie(eql)
lambda(lsl)
}
}
\seealso{
\code{\link{leslie}}
}

