% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/len2wt.R
\docType{methods}
\name{len2wt,FLQuant,FLPar-method}
\alias{len2wt}
\alias{len2wt,FLCohort,FLPar-method}
\alias{len2wt,FLQuant,FLPar-method}
\alias{len2wt,numeric,FLPar-method}
\alias{len2wt-method}
\title{Length to weight conversion}
\usage{
\S4method{len2wt}{FLQuant,FLPar}(length, params)
}
\arguments{
\item{length}{age FLQuant, FLPar or numeric}

\item{params}{\code{FLPar}}

\item{...}{any other arguments}
}
\value{
Returns a class of same type as \code{length} e.g. \code{FLQuant}
}
\description{
Converts length to weight based on $W=aL^b$
}
\examples{
\dontrun{
params=FLPar(a=1,b=3)
len2wt(FLQuant(10),params)
}
}
\seealso{
\code{\link{wt2len}}
}

