% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/len2wt.R
\docType{methods}
\name{len2wt}
\alias{len2wt}
\alias{len2wt,FLCohort,FLPar-method}
\alias{len2wt,FLQuant,FLPar-method}
\alias{len2wt,numeric,FLPar-method}
\alias{len2wt-method}
\title{len2wt}
\usage{
len2wt(length, params, ...)
}
\arguments{
\item{length}{age FLQuant, FLPar or numeric}

\item{params}{\code{FLPar}}

\item{...}{any other arguments}
}
\value{
Depends on the value of \code{length}

#'
}
\description{
converts length to weight
}
\examples{
\dontrun{
params=FLPar(a=1,b=3)
len2wt(params,FLQuant(10))
}
}
\seealso{
\code{\link{wt2len}}
}

