% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/len2wt.R
\docType{methods}
\name{len2wt}
\alias{len2wt}
\title{len2wt}
\usage{
len2wt(length, param, ...)
}
\arguments{
\item{length}{age FLQuant, FLPar or numeric}

\item{param}{}
}
\value{
Depends on the value of \code{length}

#'
}
\description{
converts length to weight
}
\examples{
\dontrun{
param=FLPar(a=1,b=3)
len2wt(param,FLQuant(10))
}
}
\seealso{
\code{\code{\link{wt2len}}}
}

