% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leslie.R
\docType{methods}
\name{leslie}
\alias{leslie}
\alias{leslie,FLBRP-method}
\alias{leslie-method}
\title{leslie}
\usage{
leslie(object, ...)
}
\arguments{
\item{object}{\code{FLBRP}}

\item{...}{any other arguments}
}
\value{
\code{matrix}
}
\description{
Creates a Leslie Matrix
}
\examples{
\dontrun{
library(ggplot2)
library(FLCore)
library(FLBRP)
library(FLife)
data(pars)
pms=lhSim(pars[[1]])
eql=lh(pms,range = c(min=0,max=8, minfbar=1,maxfbar=8,plusgroup=8))
lsl=leslie(eql,fbar=c(refpts(eql)["crash","harvest"]))
}
}
\seealso{
\code{\link{lh}}
}

