% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leslie.R
\docType{methods}
\name{leslie,FLBRP-method}
\alias{leslie,FLBRP-method}
\alias{leslie-method}
\title{leslie}
\usage{
\S4method{leslie}{FLBRP}(object, fbar = FLQuant(0), numbers = TRUE, ...)
}
\arguments{
\item{object}{\code{FLBRP}}

\item{fbar}{\code{numeric} F at whicj survival calculated}

\item{numbers}{\code{boolean} numbers or biomass, numbers bt default}

\item{...}{any other arguments}
}
\value{
\code{matrix}
}
\description{
Creates a Leslie Matrix
}
\examples{
\dontrun{
library(ggplot2)
library(FLCore)
library(FLBRP)
library(FLife)
data(pars)
pms=lhSim(pars[[1]])
eql=lh(pms,range = c(min=0,max=8, minfbar=1,maxfbar=8,plusgroup=8))
lsl=leslie(eql,fbar=c(refpts(eql)["crash","harvest"]))
}
}
\seealso{
\code{\link{lh}}
}

