% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leslie.R
\docType{methods}
\name{leslie,FLBRP-method}
\alias{leslie,FLBRP-method}
\alias{leslie-method}
\title{Leslie matrix}
\usage{
\S4method{leslie}{FLBRP}(object, fbar = FLQuant(0), numbers = TRUE, ...)
}
\arguments{
\item{object}{\code{FLBRP}}

\item{fbar}{\code{numeric} F at whicj survival calculated}

\item{numbers}{\code{boolean} numbers or biomass, numbers bt default}

\item{...}{any other arguments}
}
\value{
\code{matrix}
}
\description{
Creates a Leslie Matrix from a \code{FLBRP} object that represents a population at equilibrium
}
\examples{
\dontrun{
eql=lhEql(lhPar(FLPar(linf=100)))
leslie(eql)
}
}
\seealso{
\code{\link{r}}, \code{\link{lhRef}}, \code{\link{lhPar}}, \code{\link{lhEql}}
}

