% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhEql.R
\docType{methods}
\name{lhEql,FLPar-method}
\alias{lhEql,FLPar-method}
\alias{lhEql}
\alias{lhEql-method}
\title{Derives an \code{FLBRP} from life history parameters}
\usage{
\S4method{lhEql}{FLPar}(params, growth = FLife::vonB, m = "gislason",
  mat = logistic, sel = dnormal, sr = "bevholt", range = c(min = 0,
  max = 40, minfbar = 1, maxfbar = 40, plusgroup = 40), spwn = 0,
  fish = 0.5, units = if ("units" \%in\% names(attributes(params)))
  attributes(params)$units else NULL, midyear = 0.5, ...)
}
\arguments{
\item{params}{an \code{FLPar} object with life history parameters}

\item{growth}{function that takes an \code{FLPar} object with parameters, by default \code{vonB}}

\item{m}{\code{character} takes the natural mortality model name, by default gislason}

\item{mat}{function that takes an \code{FLPar} object with parameters, by default \code{logistic}}

\item{sel}{function that takes an \code{FLPar} object with parameters, by default \code{dnormal}}

\item{sr}{\code{character} value, "bevholt" by default}

\item{range}{\code{numeric} with age range by default from 0 to 40}

\item{spwn}{\code{numeric} give propotion of year when spawning occurs, by default is params["a50"]-floor(params["a50"])}

\item{fish}{\code{numeric} give propotion of year when fishing occurs, by default 0.5}

\item{units}{\code{character} for vectors in \code{FLBRP} returned by method}

\item{midyear}{when growth measured, default 0.5}

\item{...}{any other arguments}
}
\value{
\code{FLBRP} object
}
\description{
Takes an \code{FLPar} object with life history and selectivity parameters
and generates a corresponding \code{FLBRP} object. Can use a range of functional forms.
}
\examples{
\dontrun{
data(teleost)
alb=teleost[,"Thunnus alalunga"]
eql=lhEql(lhPar(alb))
}



}
\seealso{
\code{\link{lhPar}}, \code{\link{lhRef}}

\code{\link{vonB}} \code{\link{lorenzen}} \code{\link{sigmoid}}
}
