% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhEql.R
\docType{methods}
\name{lhEql,FLPar-method}
\alias{lhEql,FLPar-method}
\alias{lhEql}
\alias{lhEql-method}
\title{Derives an \code{FLRP} from life history parameters}
\usage{
\S4method{lhEql}{FLPar}(params, growth = FLife::vonB, m = function(length,
  params) exp(0.55) * (length^-1.61) \%*\% (params["linf"]^1.44) \%*\%
  params["k"], mat = FLife::logistic, sel = FLife::dnormal,
  sr = "bevholt", range = c(min = 0, max = 40, minfbar = 1, maxfbar = 40,
  plusgroup = 40), spwn = c(params["a50"] - floor(params["a50"])),
  fish = 0.5, units = if ("units" \%in\% names(attributes(params)))
  attributes(params)$units else NULL, midyear = 0.5, ...)
}
\arguments{
\item{params}{an \code{FLPar} object with life history parameters}

\item{growth}{A function that takes an \code{FLPar} oject with parameters, by default \code{vonB}}

\item{m}{A function that takes an \code{FLPar} oject with parameters, by default \code{gislason}}

\item{mat}{A function that takes an \code{FLPar} oject with parameters, by default \code{logistic}}

\item{sel}{A function that takes an \code{FLPar} oject with parameters, by default \code{dnormal}}

\item{sr}{A \code{character} value, "bevholt" by default}

\item{range}{A \code{numeric} with age range by default from 0 to 40}

\item{spwn}{A \code{numeric} give propotion of year when spawning occurs, by default is params["a50"]-floor(params["a50"])}

\item{fish}{A \code{numeric} give propotion of year when fishing occurs, by default 0.5}

\item{units}{A\code{character} for vectors in \code{FLRP} returned by method}

\item{midyear}{when growth measured, default 0.5}

\item{...}{any other arguments}
}
\value{
\code{FLRP} object
}
\description{
Takes an \code{FLPar} object with life history and selectivity parameters
and generates an corresponding \code{FLRP} object. Can uses a range of functional forms
}
\examples{
\dontrun{
data(teleost)
alb=teleost[,"Thunnus alalunga"]
eql=lhEql(lhPar(alb))
}

}
\seealso{
\code{\link{lhPar}}, \code{\link{lhRef}}

\code{\link{vonB}} \code{\link{lorenzen}} \code{\link{sigmoid}}
}
