% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhPar.R
\docType{methods}
\name{lhPar}
\alias{lhPar}
\title{Generates life history parameters}
\usage{
lhPar(params, t0 = function(params) -exp(-0.3922 - 0.2752 *
  log(params["linf"]) \%-\% (1.038 * log(params["k"]))), a = 3e-04, b = 3,
  ato95 = 1, sl = 1, sr = 5000, s = 0.9, v = 1000)
}
\arguments{
\item{params}{\code{FLPar} object with parameters for life history equations and selection pattern.
Need Linfinity to estimate other parameters, if any other parameters supplied in \code{code} then
these are not provided by the algorithm}

\item{t0}{of von Bertalanffy. This is a default that isnt normally derived
from life history theory, as are the following args.}

\item{a}{coefficient of length weight relationship}

\item{b}{exponent of length weight relationship}

\item{ato95}{age at which 95\% of fish are mature, offset to age at which 50\% are mature}

\item{sl}{selectivity-at-age parameter, standard deviation of lefthand limb of double normal}

\item{sr}{stock recruitment relationship}

\item{s}{steepness of stock recruitment relationship}

\item{v}{virgin biomass}
}
\value{
object of class \code{FLPar} with missing parameters calculated from life history theory
}
\description{
Uses life history theory to derive parameters for biological relationships, i.e. or growth, 
maturity, natural mortality. Selectivity by default is set so age at peak selectivity is the 
same as age at 50/% mature (a50) As a minimum all `lhPar` requires is `linf` the asymptotic 
length of the von Bertalannfy growth equation.
}
\examples{
\dontrun{
lhPar(FLPar(linf=200))
}

}
\seealso{
\code{\link{loptAge}}, \code{\link{lhRef}}, \code{\link{lhPar}}, \code{\link{lhEql}}
}
