% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/popPar.R
\docType{methods}
\name{lhRef}
\alias{lhRef}
\title{lhRef}
\usage{
lhRef(params, m = function(length, params) params["m1"] \%*\% (exp(log(length)
  \%*\% params["m2"])), sr = "bevholt", range = c(min = 0, max = 40, minfbar
  = 1, maxfbar = 40, plusgroup = 40), what = c("r", "lopt", "rc", "sk",
  "lro"))
}
\arguments{
\item{params}{\code{FLPar}}

\item{m}{function for natural mortality}
}
\value{
Depends on the value of \code{data}
}
\description{
Von Bertalanffy growth equation
}
\examples{
\dontrun{
params=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=vonB(age,params)
age=vonB(params,length=len)
}
}
\seealso{
\code{\link{gascuel}}
}

