% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhRef.R
\docType{methods}
\name{lhRef}
\alias{lhRef}
\title{lhRef}
\usage{
lhRef(params, m = function(length, params) exp(0.55) * (length^-1.61) \%*\%
  (params["linf"]^1.44) \%*\% params["k"], sr = "bevholt", range = c(min =
  0, max = 40, minfbar = 1, maxfbar = 40, plusgroup = 40), what = c("r",
  "lopt", "rc", "sk", "lro"))
}
\arguments{
\item{params}{\code{FLPar}}

\item{m}{function for natural mortality}

\item{sr}{\code{character} with stock recruitment relationship, e.g. "bevholt","ricker",...}

\item{range}{ages used i.e. c(min=0,max=40,minfbar=1,maxfbar=40,plusgroup=40)}

\item{what}{quantities to calculate "r","lopt","rc","sk","lro"}
}
\value{
Depends on the value of \code{data}
}
\description{
Von Bertalanffy growth equation
}
\examples{
\dontrun{
params=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=vonB(age,params)
age=vonB(params,length=len)
}
}
\seealso{
\code{\link{gascuel}}
}

