% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logistic.R
\docType{methods}
\name{logistic}
\alias{logistic}
\title{logistic}
\usage{
logistic(age, params, ...)
}
\arguments{
\item{age}{\code{FLQuant} with ages}

\item{params}{\code{FLPar} with parameters a50,ato95 and asym}

\item{...}{other arguments}
}
\description{
logistic function
}
\examples{
\dontrun{
params=lhPar(FLPar(linf=100))
age=FLQuant(1:10,dimnames=list(age=1:10))
mat=logistic(params,age)
}
}
\seealso{
\code{\link{gompertz}}
}

