% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\docType{methods}
\name{logistic,FLQuant,FLPar-method}
\alias{logistic}
\alias{logistic,FLQuant,FLPar-method}
\alias{logistic-method}
\title{logistic}
\usage{
\S4method{logistic}{FLQuant,FLPar}(age, params, ...)
}
\arguments{
\item{age}{\code{FLQuant} with ages}

\item{params}{\code{FLPar} with parameters a50,ato95 and asym}

\item{...}{other arguments}
}
\description{
logistic function
}
\examples{
\dontrun{
params=FLPar(a50=4,ato95=1,asym=1.0)
age=FLQuant(1:10,dimnames=list(age=1:10))
logistic(age,params)
}
}
\seealso{
\code{\link{gompertz}}
}

