% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lopt.R
\docType{methods}
\name{lopt}
\alias{lopt}
\title{lopt}
\usage{
lopt(param, ...)
}
\arguments{
\item{param}{}

\item{m}{natural mortality}

\item{eql}{}
}
\value{
FLPar with length at maximum biomass
}
\description{
Finds length at maximum biomass
}
\details{
There are several ways to calculate $L_{opt}$, i.e.
i) ${2/3}^{rds}$ of $L_{\infty}$
ii) $L_{\infty}\frac{3}{3+k/m}$
iii) by maximising the biomass of
iv) from an FLBRP object by fishing at F=0 and finding age where biomass is a maximum
}
\examples{
\dontrun{
data(pars)
lopt(pars[[1]])
}
}
\seealso{
\code{\code{\link{vonB}}}
}

