% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lopt.R
\docType{methods}
\name{lopt}
\alias{lopt}
\alias{lopt,FLPar-method}
\alias{lopt-method}
\title{lopt}
\usage{
lopt(params, ...)
}
\arguments{
\item{params}{FLPar}

\item{...}{any other arguments}
}
\value{
FLPar with length at maximum biomass
}
\description{
Finds length at maximum biomass
}
\details{
There are several ways to calculate \deqn{L_{opt}}, i.e.
i) \deqn{{2/3}^{rds}  L_{\infty}}
ii) \deqn{L_{\infty}\frac{3}{3+k/m}}
iii) by maximising the biomass of
iv) from an FLBRP object by fishing at F=0 and finding age where biomass is a maximum
}
\examples{
\dontrun{
data(pars)
lopt(pars[[1]])
}
}
\seealso{
\code{\link{vonB}}
}

