% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopt.R
\docType{methods}
\name{lopt,FLPar-method}
\alias{lopt}
\alias{lopt,FLPar-method}
\alias{lopt-method}
\title{Length at maximum biomass}
\usage{
\S4method{lopt}{FLPar}(params, m = function(length, params) exp(0.55) *
  (length^-1.61) \%*\% (params["linf"]^1.44) \%*\% params["k"], ...)
}
\arguments{
\item{params}{FLPar}

\item{m}{A function, i.e. gislason}

\item{...}{any other arguments}
}
\value{
\code{FLPar} with length at maximum biomass of a cohort
}
\description{
Finds length at maximum biomass, assuminmg natural mortality is a function of
 mass-at-age, i.e. Lorenzen.
}
\details{
There are several ways to calculate \deqn{L_{opt}}, i.e.
i) \deqn{{2/3}^{rds}  L_{\infty}}
ii) \deqn{L_{\infty}\frac{3}{3+k/m}}
iii) by maximising the biomass of
iv) from an FLBRP object by fishing at F=0 and finding age where biomass is a maximum
}
\examples{
\dontrun{
params=lhPar(FLPar(linf=100))
lopt(params)
}
}
\seealso{
\code{\link{loptAge}}, \code{\link{lhRef}}, \code{\link{lhPar}}, \code{\link{lhEql}},
}

