% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lorenzen.R
\docType{methods}
\name{lorenzen}
\alias{lorenzen}
\alias{lorenzen,FLQuant,FLPar-method}
\alias{lorenzen,FLQuant,missing-method}
\alias{lorenzen,FLQuant,numeric-method}
\title{lorenzen}
\usage{
lorenzen(wt, params, ...)
}
\arguments{
\item{wt}{mass at which M is to be predicted}

\item{params}{an \code{FLPar} with two values; i.e. a equal to M at unit mass
and b a power term; defaults are a=0.3 and b=-0.288}

\item{...}{any other arguments}
}
\description{
Lorenzen natural mortality relatoinship estimate M as a function of weight.
M=a*wt^b;
}
\examples{
\dontrun{
mass=FLQuant(c( 1.90, 4.23, 7.47,11.48,16.04,20.96,26.07,31.22,
               36.28,41.17,45.83,50.20,54.27,58.03,61.48,64.62),
             dimnames=list(age=1:16))
lorenzen(mass)
}
}
\seealso{
\code{\link{gislason}}
}

