% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/noise.R
\docType{methods}
\name{noise}
\alias{noise}
\alias{noise,numeric,FLQuant-method}
\title{noise}
\usage{
noise(n, len, ...)
}
\arguments{
\item{n}{number of iterations}

\item{len}{an \code{FLQuant}}

\item{...}{any other arguments, sd standard deviation, b autocorrelation parameter
a real number in [0,1], trunc get rid of values > abs(trunc), burn allows burn in by getting
rid of 1st values i series}
}
\value{
A vector with autocorrelation equal to B.
}
\description{
A noise generator
}
\examples{
\dontrun{
white <- noise(1000,sd=.3,b=0)
plot(white)
acf(white)

red <- noise(1000,sd=.3,b=0.8)
plot(red)
acf(white)

blue <- noise(1000,sd=.3,b=-0.8)
plot(blue)
acf(blue)

data(ple4)
res=noise(4,m(ple4),burn=10,b=0.9)

ggplot()+
geom_point(aes(year,age,size= data),
            data=subset(as.data.frame(res),data>0))+
geom_point(aes(year,age,size=-data),
            data=subset(as.data.frame(res),data<=0),colour="red")+
scale_size_area(max_size=4, guide="none")+
facet_wrap(~iter)

res=noise(4,m(ple4),burn=10,b=0.9,cohort=TRUE)
ggplot()+
geom_point(aes(year,age,size= data),
          data=subset(as.data.frame(res),data>0))+
geom_point(aes(year,age,size=-data),
          data=subset(as.data.frame(res),data<=0),colour="red")+
scale_size_area(max_size=4, guide="none")+
facet_wrap(~iter)

}
}
\references{
Ranta and Kaitala 2001 Proc. R. Soc.
vt = b * vt-1 + s * sqrt(1 - b^2)
s is normally distributed random variable with mean = 0
b is the autocorrelation parameter
}

