% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powh.R
\docType{methods}
\name{powh,numeric,numeric-method}
\alias{powh,numeric,numeric-method}
\alias{powh}
\alias{powh-method}
\title{Powell Weatherall method for estimating Z and Linfinity}
\usage{
\S4method{powh}{numeric,numeric}(len, n, weights = FALSE, fromMode = FALSE,
  linf = 0)
}
\arguments{
\item{len}{vector with length distribution}

\item{n}{vector with numbers in each length bin}

\item{weights}{weights for observations}

\item{fromMode}{booelean set to FALSE by default}

\item{linf}{allows linf to be fixed if >0}

\item{...}{any other argument, i.e. weights =TRUE, fromMode =FALSE}
}
\value{
a \code{data.frame} with \code{mn} (mean), \code{diff} (difference), 
\code{len} (length) and \code{n} (frequency)
}
\description{
Estimates growth and mortality parameters from length frequency data.
}
\details{
Beverton and Holt (1956) developed a method to estimate population parameters 
 such as total mortality (Z) from length data i.e.

\deqn{Z=K\frac{L_{infinity}-\overline{L}}{\overline{L}-L'}}
                            
Powell (1979) then developed a method, extended by Wetherall et al. (1987), 
to estimate growth and mortality parameters. This assumes that the right hand tail 
of a length frequency distribution is determined by the asymptotic length L 
and the ratio between Z and the growth rate K.

The Beverton and Holt methods assumes good estimates for K and $L_{infinity}$, 
while the Powell-Wetherall method only requires an estimate
of K, since $L_{infinity}$ is estimated by the method as well as Z/K. These method 
therefore provide estimates for Z/K, if K is unknown and 
Z if K is known.  

As well as assuming that growth follows the von Bertalanffy growth function, 
it is also assumed that the population is in a steady state 
with constant exponential mortality, no changes in selection pattern of the 
fishery and constant recruitment.

In the Powell-Wetherall method L' can take any value between the 
smallest and largest sizes. Equation 1 then provides a series of estimates of
Z and since 

\deqn{
\overline{L}-L'=a+bL'
}

a and b can be estimated by a regression analysis where 

\deqn{b={-K}/{Z+K}}
\deqn{a=-bL_{infinity}}

Therefore plotting \deqn{\overline{L}-L} against $L'$ provides an estimate 
of $L_{infinity}$ and Z/K from

\deqn{L_{infinity}=-a/b}
\deqn{Z/K={-1-b}/{b}}

If K is known Z can also be esimated
}
\examples{
\dontrun{
data(cas)
pw=ddply(subset(cas), .(year), 
  function(cas) powh(cas$len,cas$n)$data)
  
  pw=transform(pw, lustrum=(year\%/\%5)*5,
        yr    =year-(year\%/\%5)*5,
        weight=ifelse(len>=100&len<=200,1,0))
        
ggplot(pw)+
  geom_line(aes(len,diff,colour=factor(yr),group=year))+
  scale_x_continuous(limits=c(0,300)) +
  xlab("Length (cm)")+
  ylab("Difference between Length and Mean Size")+
  geom_smooth(aes(len,diff,weight=weight),
  method="lm",col="red",size=1.25,alpha=.1)+
  theme_bw()+theme(legend.position="none")
  }
}
\references{
R. Beverton and S. Holt. Review of method for estimating mortality rates in 
 exploited fish populations, with special reference to sources of bias in 
 catch sampling. Rapports et Proces-Verbaux., 140(1): 67--83, 1956.   
 
 D. G. Powell.
 Estimation of mortality and growth parameters from the length
 frequency of a catch [model].
 \emph{Rapports et Proces-Verbaux des Reunions}, 175, 1979.
 
 J. Wetherall, J. Polovina, and S. Ralston.
 Estimating growth and mortality in steady-state fish stocks from
 length-frequency data.
 \emph{ICLARM Conf. Proc}, pages 53--74, 1987.
}
\seealso{
\code{\link{cc}}
}
