% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stars.R
\name{rod}
\alias{rod}
\title{rod}
\usage{
rod(object, ...)
}
\arguments{
\item{object}{an object of class \code{FLQuant}}
}
\description{
Regime shifts
Evidence for regime shifts are explored using a a sequential t-test algorithm
(STARS; \cite{rodionov2004sequential}) as modified by Szuwalski et al., (submitted)
}
\details{
Returns a data.frame
\dontrun{
    object=rlnorm(1,FLQuant(0,dimnames=list(year=1:30)),.3)
    pg=rod(object)
    plot(object) +
       geom_polygon(aes(year,data,group=regime),
       fill="lavender",col="blue",
       lwd=.25,data=pg,alpha=.2)
}
}

