% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\docType{methods}
\name{sigmoid,FLQuant,FLPar-method}
\alias{sigmoid,FLQuant,FLPar-method}
\alias{sigmoid}
\alias{sigmoid-method}
\alias{sigmoid,FLPar,FLPar-method}
\alias{sigmoid,FLQuant,numeric-method}
\alias{sigmoid,numeric,numeric-method}
\title{Sigmoid ogive}
\usage{
\S4method{sigmoid}{FLQuant,FLPar}(age, params, ...)
}
\arguments{
\item{age}{FLQuant, FLPar or numeric with ages}

\item{params}{\code{FLPar} with age at whick probability is 0.5 (\code{a50}) and offset to 0.95 (\code{ato95})}

\item{...}{any other arguments}
}
\value{
returns an object of same type as \code{age} e.g. \code{FLQuant}
}
\description{
sigmoid ogive
}
\details{
The sigmoid ogive is an S-shaped or sigmoid curve or sigmoidal functions, 
Verhulst hypothesizes that small populations increase geometrically, because the supply of resources exceeds demand. Then, as supply and demand balance, population growth is constant. Finally, as demand exceeds supply, population growth decreases at the same rate that it had increased. Verhulst describes this process with an equation that enables him to predict when a population will reach any given size (see Verhulst's Figure):
}
\examples{
\dontrun{
params=FLPar(a50=4,ato95=1)
age=FLQuant(1:10,dimnames=list(age=1:10))
sigmoid(age,params)
}
}
\seealso{
\code{\link{knife}}, \code{\link{dnormal}}, \code{\link{logistic}}

\code{\link{dnormal},\link{knife}}
}
