% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/powh.R
\docType{methods}
\name{moment}
\alias{moment}
\alias{moment,numeric-method}
\alias{moment-method}
\title{moment}
\usage{
moment(object, ...)
}
\arguments{
\item{object}{a vector holding a time series}

\item{...}{any other arguments, i.e. x,n=rep(1,length(x)),na.rm=T}
}
\value{
a \code{vector} with the inter-annual variation each time step
}
\description{
aa
}
\examples{
x=1
}

