% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vonB.R
\docType{methods}
\name{vonB}
\alias{vonB}
\title{vonB}
\usage{
vonB(params, age, ...)
}
\arguments{
\item{params}{}

\item{age}{FLQuant, FLPar or numeric with ages}

\item{length}{FLQuant, FLPar or numeric with length, if supplied as a named paramreter
instead  of age then calculates ages.}
}
\value{
Depends on the value of \code{data}

#'
}
\description{
Von Bertalanffy growth equation
}
\examples{
\dontrun{
params=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=vonB(params,age)
age=vonB(params,length=len)
}
}
\seealso{
\code{\code{\link{gascuel}}}
}

