% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonB.R
\docType{methods}
\name{vonB,FLQuant,FLPar-method}
\alias{vonB}
\alias{vonB,FLPar,FLPar-method}
\alias{vonB,FLQuant,FLPar-method}
\alias{vonB,FLQuant,numeric-method}
\alias{vonB,missing,FLPar-method}
\alias{vonB,numeric,numeric-method}
\alias{vonB-method}
\title{von Bertalanffy growth curve}
\usage{
\S4method{vonB}{FLQuant,FLPar}(age, params, ...)
}
\arguments{
\item{age}{FLQuant, FLPar or numeric object with values corresponding to ages}

\item{params}{\code{FLPar} object with parameters \code{linf}, \code{k} and \code{t0}}

\item{...}{other arguments}
}
\value{
Returns an object of same class as \code{age}  e.g. \code{FLQuant}
}
\description{
Von Bertalanffy growth equation
}
\examples{
\dontrun{
params=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=vonB(age,params)

#inverse growth curve
vonB(params=params,length=len)
}
}
\seealso{
\code{\link{gompertz}}, \code{\link{gascuel}}
}

