% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vonB.R
\docType{methods}
\name{vonB}
\alias{vonB}
\alias{vonB,FLPar,FLPar-method}
\alias{vonB,FLQuant,FLPar-method}
\alias{vonB,FLQuant,numeric-method}
\alias{vonB,missing,FLPar-method}
\alias{vonB,numeric,numeric-method}
\title{vonB}
\usage{
vonB(age, params, ...)
}
\arguments{
\item{age}{FLQuant, FLPar or numeric object with values corresponding to ages}

\item{params}{\code{FLPar} object with parameters \code{linf}, \code{k} and \code{t0}}

\item{...}{other arguments}
}
\value{
an object the same type supplied as \code{age}
}
\description{
Von Bertalanffy growth equation
}
\examples{
\dontrun{
params=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
vonB(age,params)

#inverse growth curve
age=vonB(params,length=len)
}
}
\seealso{
\code{\link{gascuel}}
}

