% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wt2len.R
\docType{methods}
\name{wt2len}
\alias{wt2len}
\title{wt2len}
\usage{
wt2len(params, data, ...)
}
\arguments{
\item{params}{FLPar with a and b, i.e. condition and scaling factors}

\item{age}{FLQuant, FLPar or numeric with length}
}
\value{
Depends on the value of \code{data}

#'
}
\description{
converts weight to length
}
\examples{
\dontrun{
params=FLPar(a=1,b=3)
wt2len(params,FLQuant(10))
}
}
\seealso{
\code{\code{\link{len2wt}}}
}

