% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt2len.R
\docType{methods}
\name{wt2len,FLQuant,FLPar-method}
\alias{wt2len,FLQuant,FLPar-method}
\alias{wt2len}
\alias{wt2len-method}
\alias{wt2len,FLCohort,FLPar-method}
\alias{wt2len,numeric,FLPar-method}
\title{Mass to length conversion}
\usage{
\S4method{wt2len}{FLQuant,FLPar}(wt, params, ...)
}
\arguments{
\item{wt}{FLQuant, FLPar or numeric with length}

\item{params}{\code{FLPar} with a and b, i.e. condition and scaling factors}

\item{...}{any other arguments}
}
\value{
Returns an object of same class as \code{wt}  e.g. \code{FLQuant}
}
\description{
converts weight to length
}
\examples{
\dontrun{
params=FLPar(a=0.1,b=3)
wt2len(FLQuant(10),params)
}

}
\seealso{
\code{\link{len2wt}}
}
